Remove old actions
Prerelease: We should create release versions manually as we often do breaking changes. Every MC update we make to TNT it breaks any compatibility with TNT files of another version. Therefore, we should bump the version when we do this.

Gradle build: Done when publishing to the repo so unnecessary